Add Authentication

+++ demo/src/main/java/com/example/demo/services/TaskService.java
@@ -3,6 +3,7 @@ package com.example.demo.services;
 import com.example.demo.models.Task;
 import com.example.demo.repositories.TaskRepository;
 import java.util.List;
+import lombok.NonNull;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -10,7 +11,7 @@ import org.springframework.stereotype.Service;
 public class TaskService {
   @Autowired private TaskRepository taskRepository;
 
-  public Task createNewTask(Task task) {
+  public Task createNewTask(@NonNull Task task) {
     return taskRepository.save(task);
   }
 
@@ -32,10 +33,13 @@ public class TaskService {
 
   public void deleteTask(Long id) {
     Task task = taskRepository.getById(id);
+    if (task == null) {
+      return;
+    }
     taskRepository.delete(task);
   }
 
-  public Task updateTask(Task task) {
+  public Task updateTask(@NonNull Task task) {
     return taskRepository.save(task);
   }
 }
